+++ src/include/anopa/service.h
@@ -2,7 +2,7 @@
  * anopa - Copyright (C) 2015-2017 Olivier Brunel
  *
  * service.h
- * Copyright (C) 2015 Olivier Brunel <jjk@jjacky.com>
+ * Copyright (C) 2015-2017 Olivier Brunel <jjk@jjacky.com>
  *
  * This file is part of anopa.
  *
@@ -23,12 +23,12 @@
 #ifndef AA_SERVICE_H
 #define AA_SERVICE_H
 
+#include <stdint.h>
+#include <sys/types.h>
 #include <signal.h> /* pid_t */
 #include <skalibs/stralloc.h>
 #include <skalibs/genalloc.h>
-#include <skalibs/uint16.h>
 #include <skalibs/tai.h>
-#include <s6/ftrigr.h>
 #include <anopa/service_status.h>
 
 #define AA_START_FILENAME           "start"
@@ -40,7 +40,7 @@ extern stralloc aa_names;
 extern genalloc aa_main_list;
 extern genalloc aa_tmp_list;
 extern genalloc aa_pid_list;
-extern int aa_secs_timeout;
+extern unsigned int aa_secs_timeout;
 
 #define aa_service(i)               (&((aa_service *) aa_services.s)[i])
 #define aa_service_name(service)    (aa_names.s + (service)->offset_name)
@@ -77,17 +77,17 @@ typedef enum
 
 typedef struct
 {
-    int offset_name;
+    size_t offset_name;
     int nb_mark;
     genalloc needs;
     genalloc wants;
     genalloc after;
-    int secs_timeout;
+    unsigned int secs_timeout;
     aa_ls ls;
     aa_service_status st;
     tain_t ts_exec;
     /* longrun */
-    uint16 ft_id;
+    uint16_t ft_id;
     int gets_ready;
     /* oneshot */
     int fd_in;
@@ -100,21 +100,21 @@ typedef struct
 
 typedef void (*aa_close_fd_fn) (int fd);
 typedef void (*aa_autoload_cb) (int si, aa_al al, const char *name, int err);
-typedef void (*aa_prepare_cb) (int si, int si_next, int is_needs, int first);
+typedef void (*aa_prepare_cb) (int si, int si_next, int is_needs, size_t first);
 typedef void (*aa_scan_cb) (int si, int sni);
 typedef void (*aa_exec_cb) (int si, aa_evt evt, pid_t pid);
 
-extern void aa_free_services (aa_close_fd_fn close_fd_fn);
-extern int  aa_add_name (const char *name);
-extern int  aa_get_service (const char *name, int *si, int new_in_main);
-extern void aa_unmark_service (int si);
-extern int  aa_mark_service (aa_mode mode, int si, int in_main, int no_wants, aa_autoload_cb al_cb);
-extern int  aa_preload_service (int si);
-extern int  aa_ensure_service_loaded (int si, aa_mode mode, int no_wants, aa_autoload_cb al_cb);
-extern int  aa_prepare_mainlist (aa_prepare_cb prepare_cb, aa_exec_cb exec_cb);
-extern void aa_scan_mainlist (aa_scan_cb scan_cb, aa_mode mode);
-extern int  aa_exec_service (int si, aa_mode mode);
-extern int  aa_get_longrun_info (uint16 *id, char *event);
-extern int  aa_unsubscribe_for (uint16 id);
+extern void     aa_free_services (aa_close_fd_fn close_fd_fn);
+extern size_t   aa_add_name (const char *name);
+extern int      aa_get_service (const char *name, int *si, int new_in_main);
+extern void     aa_unmark_service (int si);
+extern int      aa_mark_service (aa_mode mode, int si, int in_main, int no_wants, aa_autoload_cb al_cb);
+extern int      aa_preload_service (int si);
+extern int      aa_ensure_service_loaded (int si, aa_mode mode, int no_wants, aa_autoload_cb al_cb);
+extern int      aa_prepare_mainlist (aa_prepare_cb prepare_cb, aa_exec_cb exec_cb);
+extern void     aa_scan_mainlist (aa_scan_cb scan_cb, aa_mode mode);
+extern int      aa_exec_service (int si, aa_mode mode);
+extern int      aa_get_longrun_info (uint16_t *id, char *event);
+extern int      aa_unsubscribe_for (uint16_t id);
 
 #endif /* AA_SERVICE_H */
